backend/requirements.txt
Django>=6.0.3
djangorestframework>=3.17.1
django-cors-headers>=4.9.0
PyMySQL>=1.1.2
Pillow>=12.1.1
